pyproject.toml
[tool.poetry]
name = "aisuite"
version = "0.1.6"
description = "Uniform access layer for LLMs"
authors = ["Andrew Ng"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
anthropic = { version = "^0.30.1", optional = true }
boto3 = { version = "^1.34.144", optional = true }
vertexai = { version = "^1.63.0", optional = true }
groq = { version = "^0.9.0", optional = true }
mistralai = { version = "^1.0.3", optional = true }
openai = { version = "^1.35.8", optional = true }
#rerankers = { version = "0.8.0"}
#google-cloud-discoveryengine = {version = "0.13.6" }

# Optional dependencies for different providers
google-cloud-discoveryengine = "^0.13.6"
rerankers = "^0.8.0"
[tool.poetry.extras]
anthropic = ["anthropic"]
aws = ["boto3"]
azure = []
google = ["vertexai"]
groq = ["groq"]
huggingface = []
mistral = ["mistralai"]
ollama = []
openai = ["openai"]
all = ["anthropic", "aws", "google", "groq", "mistral", "openai"]  # To install all providers

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pre-commit = "^3.7.1"
black = "^24.4.2"
python-dotenv = "^1.0.1"
openai = "^1.35.8"
groq = "^0.9.0"
anthropic = "^0.30.1"
notebook = "^7.2.1"
ollama = "^0.2.1"
mistralai = "^1.0.3"
boto3 = "^1.34.144"
fireworks-ai = "^0.14.0"
chromadb = "^0.5.4"
sentence-transformers = "^3.0.1"
datasets = "^2.20.0"
vertexai = "^1.63.0"
google-generativeai = "0.8.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
